procurment-tool/requirements.txt
streamlit>=1.47.1
pandas>=2.3.1
openpyxl>=3.1.5
pdf2image>=1.17.0
pdfplumber>=0.11.7
pillow>=11.3.0
pytesseract>=0.3.13
python-docx>=1.2.0
supabase>=2.0.0
python-dotenv>=1.0.0